+++ CHANGELOG.md
@@ -7,6 +7,6 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- 🎉 Initial release adding the `Tag` type.
+- 🎉 Initial release adding the `Tag` and `TagUnion` types.
 
 [0.1.0]: https://github.com/sunsided/just-a-tag/releases/tag/0.1.0